refactor(ts): extract timeline/simulation.ts from store
Move the compile -> simulate -> project pipeline (compiledScenario,
compiledTimeline, simulation, optimizerProjection and all projected view
models: logs, series, effect layouts, combo windows, requisite warnings,
gauge series, timeContext, globalExtensions) into a useTimelineSimulation(deps)
composable, mirroring the optimizer. This isolates the adapter call sites that
Phase 3 will dissolve. Golden + simulator suites unchanged. No behavior change.
